Removed situational zsh aliases

+++ .zshrc
@@ -110,11 +110,20 @@ alias adb-forward-v2rayng-waydroid="~/Android/Sdk/platform-tools/adb -s 192.168.
 alias avd-unlock="rm .android/avd/Resizable_Experimental_API_33.avd/*.lock"
 alias avd-unlock="rm .android/avd/Resizable_Experimental_API_33.avd/*.lock"
 alias waydroid-adb-connect="adb connect 192.168.240.112:5555"
+
+# Legion conservation mode
+if [ -e "/sys/devices/pci0000:00/0000:00:14.3/PNP0C09:00/VPC2004:00/conservation_mode" ]; then
 alias conservation-mode-on="echo 1 | sudo tee /sys/devices/pci0000:00/0000:00:14.3/PNP0C09:00/VPC2004:00/conservation_mode 1>/dev/null"
 alias conservation-mode-off="echo 0 | sudo tee /sys/devices/pci0000:00/0000:00:14.3/PNP0C09:00/VPC2004:00/conservation_mode 1>/dev/null"
+fi
+
 alias lotfan="sudo"
 alias "gitmain"="git checkout -f main"
+
+# Sharif net2 login
+if [ -e "~/.net2" ]; then
 source ~/.net2
+fi
 
 #THIS MUST BE AT THE END OF THE FILE FOR SDKMAN TO WORK!!!
 export SDKMAN_DIR="$HOME/.sdkman"
